Move files to features/ to simplify structure
