+++ src/features/certificate-authorities/main/request-system-cas.injectable.linux.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { requestSystemCAsInjectionToken } from "./request-system-cas-token";
+import { requestSystemCAsInjectionToken } from "../common/request-system-cas-token";
 
 const requestSystemCAsInjectable = getInjectable({
   id: "request-system-cas",
